pyproject.toml
[project]
name = "technical-analysis-mcp"
version = "0.1.0"
description = "MCP server for technical analysis using yfinance"
authors = [{ name = "Alejandro Claro" }]
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
  "fastmcp>=2.14.3",
  "pandas>=2.3.3",
  "yfinance>=1.0",
]

[dependency-groups]
dev = [
  "PyHamcrest>=2.1.0",
  "conventional-commits-check>=2.9.0",
  "pymarkdownlnt>=0.9.34",
  "pyright>=1.1.408",
  "pytest-asyncio>=1.3.0",
  "pytest-cov>=5.0.0",
  "pytest>=9.0.2",
  "ruff>=0.14.13",
]

[project.scripts]
server = "technical_analysis_mcp.server.server:main"
repl = "technical_analysis_mcp.repl.repl:main"

[tool.semantic_release]
branches = ["main"]
version_toml = ["pyproject.toml:project.version"]
version_variable = ["src/technical_analysis_mcp/__init__.py:__version__"]
commit_version_number = true
commit_subject = "chore(release): {version}"

[tool.semantic_release.remote]
type = "github"
token = "${env.GITHUB_TOKEN}"

[tool.semantic_release.publish]
dist_glob_patterns = ["dist/*"]
upload_to_vcs_release = true

[tool.pytest]
pythonpath = ["src"]
addopts = ["--cov", "--cov-report=term", "--cov-report=xml"]

[tool.coverage.run]
source = ["src"]
omit = ["tests/*"]
branch = true

[tool.coverage.report]
show_missing = true
skip_covered = false
fail_under = 80

[tool.coverage.xml]
output = "reports/coverage.xml"

[tool.pyright]
venvPath = "."
venv = ".venv"

exclude = [
  ".git",
  ".venv",
  "__pycache__",
  "build",
  "dist"
]

[tool.ruff]
line-length = 120

include = [
  "pyproject.toml",
  "src/**/*.py",
  "tests/*.py",
  "tests/**/*.py"
]

exclude = [
  ".git",
  ".venv",
  "__pycache__",
  "build",
  "dist"
]

[tool.ruff.lint]
select = ["ALL"]

ignore = [
  "COM812",
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.mccabe]
max-complexity = 10

[tool.pymarkdown]
mode.strict-config = true
plugins.md013.code_blocks = false
